Fixing segfault
This was due to the missing final NULL in he keyword list. This should have
always have been there, but until CPython 3.6 the error wasn't evident (and
even with 3.6 it's not always reproducible -- for instance, the unit tests all
nicely run).

I'm also extending the python versions used by Travis to include 3.5 and 3.6
for better coverage.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>